update store setter helpers
